move ModSettingsContainer to seperate component
